seed script only deletes unowned fruits

+++ models/seed.js
@@ -51,7 +51,8 @@ db.on('open', () => {
         { name: 'Coconut', color: 'brown', readyToEat: true }
     ]
     // then we delete every fruit in the database(all instances of this resource)
-    Fruit.deleteMany({})
+    // this will delete any fruits that are not owned by a user
+    Fruit.deleteMany({ owner: null })
         .then(() => {
             // then we'll seed(create) our starter fruits
             Fruit.create(startFruits)
